requirements.txt
azure-core==1.34.0
azure-ai-documentintelligence==1.0.2
instructor==1.10.0
numpy==2.2.6
openai==1.92.2
pydantic==2.11.5
RapidFuzz==3.13.0
requests==2.32.3
scipy==1.15.3
httpx==0.28.1
